chore: rename files in cache module
