+++ hipcheck/src/cache/mod.rs
@@ -1,4 +1,4 @@
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod plugin_cache;
-pub mod repo_cache;
+pub mod plugin;
+pub mod repo;
