+++ hipcheck/src/cache/plugin.rs
